pyproject.toml
[project]
name = "achmedius-ai"
version = "0.1.0"
description = "Automate viral video topic and visual prompt generation."
authors = [
    { name = "Your Name", email = "your@email.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.77.0",
    "requests",
    "python-dotenv"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["lib"]

[tool.pdm]
# PDM-specific settings can go here

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.8.0",
    "isort>=5.13.2",
]
